refactor: update data-service costs to use generic components from shared libs
