--- apps/data-service-catalog/components/data-service-form/components/costs-modal/costs-modal.module.scss
@@ -1,18 +1,0 @@
-.padding {
-  padding-bottom: 0.5rem;
-}
-
-.modalContent {
-  margin: var(--ds-size-4) 0 var(--ds-size-4) 0;
-  padding: var(--ds-size-6);
-  border: 1px solid var(--ds-color-neutral-border-subtle);
-  border-radius: var(--ds-border-radius-md);
-  gap: 1rem;
-}
-
-.valueCurrencyFieldset {
-  display: flex;
-  flex-direction: row;
-  gap: 1rem;
-  margin-bottom: 1rem;
-}
