+++ apps/data-service-catalog/components/details-page-columns/details-page-left-column.tsx
@@ -4,7 +4,7 @@ import {
   ISOLanguage,
 } from "@catalog-frontend/types";
 import styles from "./details-columns.module.css";
-import { InfoCard } from "@catalog-frontend/ui";
+import { CostList, InfoCard } from "@catalog-frontend/ui";
 import { isEmpty } from "lodash";
 import { localization, getTranslateText } from "@catalog-frontend/utils";
 import { Paragraph, Tag } from "@digdir/designsystemet-react";
@@ -12,7 +12,6 @@ import { DetailsUrlList } from "./components/details-url-list";
 import { ReferenceDataTag } from "./components/reference-data-tag";
 import { FormatList } from "./components/format-list";
 import { DatasetList } from "./components/dataset-list";
-import { CostList } from "./components/cost-list";
 
 type Props = {
   dataService: DataService;
@@ -115,7 +114,7 @@ export const LeftColumn = ({
 
       {!isEmpty(dataService?.costs) && (
         <InfoCard.Item
-          title={localization.dataServiceForm.fieldLabel.costs}
+          title={localization.cost.fieldLabel.costs}
           data-testid="data-service-costs"
         >
           <CostList costs={dataService.costs} language={language} />
