+++ libs/types/src/lib/cost.ts
@@ -1,0 +1,8 @@
+import { LocalizedStrings } from "./localization";
+
+export type Cost = {
+  value?: number;
+  description?: LocalizedStrings;
+  documentation?: string[];
+  currency?: string;
+};
