+++ libs/ui/src/lib/costs/cost-list.tsx
@@ -1,26 +1,23 @@
-import { DataServiceCost } from "@catalog-frontend/types";
+import { Cost } from "@catalog-frontend/types";
 import { Card, Link, List, Paragraph } from "@digdir/designsystemet-react";
-import React from "react";
 import { getTranslateText } from "@catalog-frontend/utils";
+import styles from "./costs.module.scss";
 
 type Props = {
-  costs: DataServiceCost[] | undefined;
-  language: string;
+  costs: Cost[] | undefined;
+  language?: string;
 };
 
 export const CostList = ({ costs, language }: Props) => {
+  if (!costs?.length) return null;
+
   return (
     <>
-      {costs?.map((cost, i) => (
+      {costs.map((cost, i) => (
         <Card key={`costs-card-${i}`} data-color="neutral">
           <Card.Block>
-            <List.Unordered
-              style={{
-                listStyle: "none",
-                paddingLeft: 0,
-              }}
-            >
-              {cost?.value && (
+            <List.Unordered className={styles.list}>
+              {cost?.value !== undefined && cost?.value !== null && (
                 <List.Item>
                   {cost.value} {cost.currency?.split("/")?.reverse()[0] ?? ""}
                 </List.Item>
