+++ libs/ui/src/lib/costs/costs-table.tsx
@@ -1,36 +1,45 @@
-import {
-  DataService,
-  DataServiceCost,
-  ISOLanguage,
-  ReferenceDataCode,
-} from "@catalog-frontend/types";
+"use client";
+
+import { Cost, ISOLanguage, ReferenceDataCode } from "@catalog-frontend/types";
 import { getTranslateText, localization } from "@catalog-frontend/utils";
-import {
-  Button,
-  Card,
-  Link,
-  List,
-  Paragraph,
-  Tag,
-} from "@digdir/designsystemet-react";
+import { Card, Link, List, Paragraph, Tag } from "@digdir/designsystemet-react";
 import { useFormikContext } from "formik";
-import { TitleWithHelpTextAndTag } from "@catalog-frontend/ui";
-import styles from "../data-service-form.module.css";
-import { TrashIcon } from "@navikt/aksel-icons";
-import { CostsModal } from "./costs-modal/costs-modal";
+import { DeleteButton } from "../button";
+import { TitleWithHelpTextAndTag } from "../title-with-help-text-and-tag";
+import { CostsModal } from "./costs-modal";
+import styles from "./costs.module.scss";
+
+type CostsFormValues = {
+  costs?: Cost[];
+};
 
 type Props = {
   currencies?: ReferenceDataCode[];
+  helpText?: string;
 };
 
+const CURRENCY_PRIORITY = [
+  "JPY",
+  "ISK",
+  "SEK",
+  "DKK",
+  "GBP",
+  "USD",
+  "EUR",
+  "NOK",
+];
+
+const ALLOWED_LANGUAGES = Object.freeze<ISOLanguage[]>(["nb", "nn", "en"]);
+
 const sortCurrencies = (currencies?: ReferenceDataCode[]) => {
-  const priority = ["JPY", "ISK", "SEK", "DKK", "GBP", "USD", "EUR", "NOK"];
-  return currencies?.sort((a, b) => {
+  if (!currencies) return currencies;
+
+  return [...currencies].sort((a, b) => {
     if (!a.code) return 1;
     if (!b.code) return -1;
 
-    const indexA = priority.indexOf(a.code);
-    const indexB = priority.indexOf(b.code);
+    const indexA = CURRENCY_PRIORITY.indexOf(a.code);
+    const indexB = CURRENCY_PRIORITY.indexOf(b.code);
 
     if (indexA !== -1 || indexB !== -1) {
       return indexB - indexA;
@@ -40,10 +49,9 @@ const sortCurrencies = (currencies?: ReferenceDataCode[]) => {
   });
 };
 
-export const CostsTable = ({ currencies }: Props) => {
-  const { values, setFieldValue } = useFormikContext<DataService>();
+export const CostsTable = ({ currencies, helpText }: Props) => {
+  const { values, setFieldValue } = useFormikContext<CostsFormValues>();
   const sortedCurrencies = sortCurrencies(currencies);
-  const allowedLanguages = Object.freeze<ISOLanguage[]>(["nb", "nn", "en"]);
 
   const handleDeleteCost = (index: number) => () => {
     const newCosts = values.costs?.filter((_, i) => i !== index);
@@ -53,20 +61,16 @@ export const CostsTable = ({ currencies }: Props) => {
   return (
     <div className={styles.fieldContainer}>
       <TitleWithHelpTextAndTag
-        helpText={localization.dataServiceForm.helptext.costs}
+        helpText={helpText ?? localization.cost.helptext.costs}
       >
-        {localization.dataServiceForm.fieldLabel.costs}
+        {localization.cost.fieldLabel.costs}
       </TitleWithHelpTextAndTag>
+
       {values?.costs?.map((item, i) => (
         <Card key={`costs-card-${i}`} data-color="neutral">
           <Card.Block className={styles.costContent}>
-            <List.Unordered
-              style={{
-                listStyle: "none",
-                paddingLeft: 0,
-              }}
-            >
-              {item?.value && (
+            <List.Unordered className={styles.list}>
+              {item?.value !== undefined && item?.value !== null && (
                 <List.Item>
                   {item.value} {item.currency?.split("/")?.reverse()[0] ?? ""}
                 </List.Item>
@@ -85,38 +89,26 @@ export const CostsTable = ({ currencies }: Props) => {
                 template={item}
                 type="edit"
                 currencies={sortedCurrencies}
-                onSuccess={(updatedItem: DataServiceCost) =>
+                onSuccess={(updatedItem: Cost) =>
                   setFieldValue(`costs[${i}]`, updatedItem)
                 }
               />
 
-              <Button
-                variant="tertiary"
-                data-color="danger"
-                onClick={handleDeleteCost(i)}
-              >
-                <TrashIcon title="Slett" fontSize="1.5rem" />
-                {localization.button.delete}
-              </Button>
+              <DeleteButton onClick={handleDeleteCost(i)} />
             </div>
           </Card.Block>
           <Card.Block className={styles.costFooter}>
             <Paragraph>{getTranslateText(item?.description)}</Paragraph>
             <div>
-              {allowedLanguages
-                .filter(
-                  (lang) =>
-                    item?.description &&
-                    Object.prototype.hasOwnProperty.call(
-                      item.description,
-                      lang,
-                    ),
-                )
-                .map((lang) => (
-                  <Tag key={lang} data-color="third">
-                    {localization.language[lang]}
-                  </Tag>
-                ))}
+              {ALLOWED_LANGUAGES.filter(
+                (lang) =>
+                  item?.description &&
+                  Object.prototype.hasOwnProperty.call(item.description, lang),
+              ).map((lang) => (
+                <Tag key={lang} data-color="info">
+                  {localization.language[lang]}
+                </Tag>
+              ))}
             </div>
           </Card.Block>
         </Card>
@@ -129,8 +121,8 @@ export const CostsTable = ({ currencies }: Props) => {
           currencies={sortedCurrencies}
           onSuccess={(formValues) =>
             setFieldValue(
-              values.costs && values?.costs.length > 0
-                ? `costs[${values?.costs?.length}]`
+              values.costs && values.costs.length > 0
+                ? `costs[${values.costs.length}]`
                 : "costs[0]",
               formValues,
             )
